// web/src/shared/api/generated/model/updateQuestionRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { JsonElement } from './jsonElement';
import type { QuestionDifficulty } from './questionDifficulty';
import type { QuestionType } from './questionType';

export interface UpdateQuestionRequest {
  type: null | QuestionType;
  /** @nullable */
  stem: string | null;
  body: JsonElement;
  gradingSpec: JsonElement;
  /** @nullable */
  explanation: string | null;
  difficulty: null | QuestionDifficulty;
  /** @nullable */
  objectiveId: string | null;
  /** @nullable */
  tags: string[] | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  maxScore: number | string | null;
}